Feature/ticket

+++ src/main/java/qnt/moviebooking/entity/ShowtimeEntity.java
@@ -8,6 +8,7 @@ import lombok.NoArgsConstructor;
 import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.UpdateTimestamp;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
@@ -21,7 +22,7 @@ public class ShowtimeEntity {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     private LocalDateTime startTime;
-    private Integer basePrice;
+    private BigDecimal basePrice;
     @Column(updatable = false)
     @CreationTimestamp
     private LocalDateTime createdAt;
